Add FFI for CachedOp Invocation
