+++ python/mxnet/cython/base.pyi
@@ -114,14 +114,6 @@ cdef extern from "mxnet/c_api.h":
                           CachedOpHandle *out,
                           _bool thread_safe);
     int MXFreeCachedOp(CachedOpHandle handle);
-    int MXInvokeCachedOp(CachedOpHandle handle,
-                         int num_inputs,
-                         NDArrayHandle *inputs,
-                         int default_ctx_type,
-                         int default_ctx_dev_id,
-                         int *num_outputs,
-                         NDArrayHandle **outputs,
-                         const int **out_stypes);
     int MXCachedOpRegisterOpHook(NDArrayHandle handle,
                                  CachedOpMonitorCallback callback,
                                  _bool monitor_all);
